Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,7 +4,7 @@ int main() {
     int radius;
     scanf("%d", &radius);
     float area;
-    area= ((3.14)*((radius)(radius)));
+    area= ((3.14)*((float)(radius)(radius)));
     printf("Area: %f",area);
     return 0;
 }
